requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
libs==0.0.10
multidict==6.0.5
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
yarl==1.9.4
